refactor(common): add `define_test_logger` macro for unified `TestLogger` creation
Also move this new and the existing test logging utilities to a
`test::logging` module.
