+++ mithril-client/src/aggregator_client.rs
@@ -877,7 +877,7 @@ mod tests {
     mod warn_if_api_version_mismatch {
         use http::response::Builder as HttpResponseBuilder;
 
-        use mithril_common::test::MemoryDrainForTestInspector;
+        use mithril_common::test::logging::MemoryDrainForTestInspector;
 
         use super::*;
