+++ mithril-common/src/test/logging/memory_logger.rs
